+++ fastn-core/src/auth/ud.rs
@@ -1,8 +1,1 @@
-#[derive(serde::Deserialize, Clone)]
-pub struct UserData {
-    pub id: i64,
-    pub username: String,
-    pub name: String,
-    pub email: String,
-    pub verified_email: bool,
-}
+pub use fastn_ds::ud::UserData;
