+++ fastn-ds/Cargo.toml
@@ -21,3 +21,8 @@ reqwest.workspace = true
 once_cell.workspace = true
 url.workspace = true
 regex.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+mime_guess.workspace = true
+colored.workspace = true
+diesel.workspace = true
