+++ fastn-ds/src/ud.rs
@@ -1,0 +1,8 @@
+#[derive(serde::Deserialize, Clone)]
+pub struct UserData {
+    pub id: i64,
+    pub username: String,
+    pub name: String,
+    pub email: String,
+    pub verified_email: bool,
+}
